backend/requirements.txt
absl-py==2.0.0
albucore==0.0.6
albumentations==1.4.8
annotated-types==0.7.0
ansible-base==2.10.8
anyio==3.7.1
apache-libcloud==3.2.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.20.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.1.0
blinker==1.8.2
boto3==1.34.49
botocore==1.34.49
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.1
certifi==2020.6.20
cffi==1.16.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
collab==1.0.198
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configparser==6.0.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.2
fastapi==0.99.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==23.5.26
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
gast==0.5.4
gattlib==0.20201113
gdown==5.1.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.13.0
h5py==3.9.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httptools==0.1.2
httpx==0.27.0
identify==2.5.33
idna==3.3
imageio==2.34.1
importlib-metadata==4.6.4
install==1.3.5
ipykernel==6.29.3
ipython==7.31.1
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kazam==1.4.5
keras==2.14.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.3
meld==3.20.4
minio==5.0.10
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==0.0.0
msal==1.28.0
msal-extensions==1.1.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
nodeenv==1.8.0
notebook==7.1.2
notebook_shim==0.2.4
ntlm-auth==1.4.0
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.1.0
ply==3.11
portalocker==2.8.2
pre-commit==3.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.24.3
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyBluez==0.23
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.11.2
PyGObject==3.42.1
pygraphviz==1.7
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
qrcode==6.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
reportlab==3.6.8
requests==2.31.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.0
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.13.1
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==3.3.1
selinux==3.3
Send2Trash==1.8.2
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post10
sniffio==1.3.0
sns==0.1
soupsieve==2.3.1
ssh-import-id==5.11
starlette==0.27.0
sympy==1.9
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.18.1
texttable==1.6.4
threadpoolctl==3.2.0
tifffile==2024.5.22
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.13.0
uvloop==0.19.0
virtualenv==20.25.0
wadllib==1.3.6
watchgod==0.6
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
websockets==8.1
Werkzeug==3.0.0
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.14.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
